[ci][bisect/4] move bisect triggering to state machine base class
