+++ release/ray_release/test_automation/state_machine.py
@@ -1,5 +1,6 @@
 import abc
 from typing import List
+from datetime import datetime, timedelta
 
 import github
 from github import Github
@@ -28,6 +29,7 @@ TEAM = [
     "serve",
     "serverless",
 ]
+MAX_BISECT_PER_DAY = 10  # Max number of bisects to run per day for all tests
 
 
 class TestStateMachine(abc.ABC):
@@ -276,5 +278,52 @@ class TestStateMachine(abc.ABC):
             f"found by bisect job https://buildkite.com/{BUILDKITE_ORGANIZATION}/"
             f"{BUILDKITE_BISECT_PIPELINE}/builds/{bisect_build_number}"
         )
-
         return True
+
+    def _trigger_bisect(self) -> None:
+        if self._bisect_rate_limit_exceeded():
+            logger.info(f"Skip bisect {self.test.get_name()} due to rate limit")
+            return
+        build = self.ray_buildkite.builds().create_build(
+            BUILDKITE_ORGANIZATION,
+            BUILDKITE_BISECT_PIPELINE,
+            "HEAD",
+            "master",
+            message=f"[ray-test-bot] {self.test.get_name()} failing",
+            env={
+                "UPDATE_TEST_STATE_MACHINE": "1",
+            },
+        )
+        failing_commit = self.test_results[0].commit
+        passing_commits = [r.commit for r in self.test_results if r.is_passing()]
+        if not passing_commits:
+            logger.info(f"Skip bisect {self.test.get_name()} due to no passing commit")
+            return
+        passing_commit = passing_commits[0]
+        self.ray_buildkite.jobs().unblock_job(
+            BUILDKITE_ORGANIZATION,
+            BUILDKITE_BISECT_PIPELINE,
+            build["number"],
+            build["jobs"][0]["id"],  # first job is the blocked job
+            fields={
+                "test-name": self.test.get_name(),
+                "passing-commit": passing_commit,
+                "failing-commit": failing_commit,
+                "concurrency": "3",
+                "run-per-commit": "1",
+                "test-type": "release-test",
+            },
+        )
+        self.test[Test.KEY_BISECT_BUILD_NUMBER] = build["number"]
+
+    def _bisect_rate_limit_exceeded(self) -> bool:
+        """
+        Check if we have exceeded the rate limit of bisects per day.
+        """
+        builds = self.ray_buildkite.builds().list_all_for_pipeline(
+            BUILDKITE_ORGANIZATION,
+            BUILDKITE_BISECT_PIPELINE,
+            created_from=datetime.now() - timedelta(days=1),
+            branch="master",
+        )
+        return len(builds) >= MAX_BISECT_PER_DAY
